Re-enable some MwType2AdvancedTests
These tests fail so I had to make sure that the assertions occur after
the threads are joined, so that all the cleanup still happens and the
test run doesn't abort.

The comment that the tests fail on (Mellanox) hardware doesn't seem to
be a reason not to run them as we have other tests that fail there too.