Differentiate MrOnOtherPd and MrMwOnOtherPd
As far as I could tell, MrOnOtherPd should use the QP's PD for the MW,
but it was using the alternate PD instead, just like MrMwOnOtherPd. Thus
the two tests looked like they were running the same test with slightly
different failure criterion. Unfortunately, though, when I apply this
fix to the test, Mellanox starts failing the test. I am going to put
this down to oversight both in the test set, and in any use Mellanox
make of the test set.