PS-12949: add short block waiting for completion channels in the comp_channel tests
The completion channel tests poll the CQ for new CQEs and then check to
see if the completion channel for that CQ is ready. At the moment there
is a race between when the driver marks the channel as ready and when the
test checks it is ready causing random test failures.

This commit makes the select block for a short period of time to allow for the delay
between a CQE becoming visible in the CQ via polling and the completion channel
becoming readable.